Don't create large intermediary tensors in the backward of matmul
Currently, if we multiply a transposed batch of matrices with shape
[b, m, n] and a matrix with shape [n, k], when computing the gradient
of the matrix, we instantiate a matrix of shape [b, m, n]. If the matrix
is large, creating this unnecessary batch of matrices may be time and
memory consuming. In this case, we fold the batch of matrices into a
matrix, which avoids creating any large intermediary tensor.

Note that multiplying a batch of matrices and a matrix naturally occurs
within an attention module, so this case surely happens in the wild.

ghstack-source-id: 558c9563841dba45bbafcd6acf1c5709f3329deb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/95261